+++ test/unit/gen_class/timeformat_get_hack.py
@@ -32,6 +32,7 @@ import version
 
 __version__ = version.__version__
 
+
 class UnitTest(unittest.TestCase):
 
     """Class:  UnitTest