+++ test/unit/gen_class/timeformat_init.py
@@ -33,6 +33,7 @@ import version
 
 __version__ = version.__version__
 
+
 class UnitTest(unittest.TestCase):
 
     """Class:  UnitTest